+++ src/comms/PCIe/TRXLooper_PCIE.cpp
@@ -82,10 +82,10 @@ class AvgRmsCounter
 
 static inline int64_t ts_to_us(int64_t fs, int64_t ts)
 {
-    int n, r;
+    int64_t n, r;
     n = (ts / fs);
     r = (ts % fs);
-    return static_cast<int64_t>(n) * 1000000 + ((static_cast<int64_t>(r) * 1000000) / fs);
+    return n * 1000000 + (r * 1000000 / fs);
 }
 
 TRXLooper_PCIE::TRXLooper_PCIE(
